Fixed deprecations

+++ src/View/Helper/Number/Locale.php
@@ -55,6 +55,6 @@ class Locale
 	public function transform( $number, ?int $decimals = null ) : string
 	{
 		$this->formatter->setAttribute( \NumberFormatter::FRACTION_DIGITS, $decimals !== null ? (int) $decimals : 2 );
-		return $this->formatter->format( (double) $number );
+		return $this->formatter->format( (float) $number );
 	}
 }
